requirements.txt
h5py==3.11.0
imageio==2.34.1
matplotlib==3.8.0
numpy==1.23.5
pandas==1.5.3
Pillow==10.4.0
scikit_learn==1.2.1
torch==2.4.0
torchvision==0.15.1
tqdm==4.66.5
